# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ChefMate",
        "description": "API para gestión de usuarios, recetas, listas y comentarios de ChefMate.",
        "contact": {
            "name": "Francisco Javier Cabello Rueda"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local de desarrollo"
        },
        {
            "url": "http://chefmate-internal.duckdns.org/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/usuarios": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "09dec89b10edba6408b0a395a96aadeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellidos",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Laura"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Martínez Ruiz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "laura@chefmate.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secreto123"
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/usuario": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener los datos del usuario autenticado",
                "operationId": "002c1a91784d76ec38db18f429e08623",
                "responses": {
                    "200": {
                        "description": "Datos del usuario obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar el perfil del usuario autenticado",
                "operationId": "2c7d747af08938cc4d90dbf49fbcdd49",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Pérez García"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nuevoPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar la cuenta del usuario autenticado",
                "operationId": "973602edc27fa66105f64e1085f61285",
                "responses": {
                    "200": {
                        "description": "Cuenta eliminada correctamente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/usuario/foto": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar la foto del usuario autenticado",
                "operationId": "134691773dd536b251bb387a8e48208d",
                "responses": {
                    "200": {
                        "description": "Foto mostrada correctamente"
                    },
                    "404": {
                        "description": "Foto no disponible"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de acceso obtenido en /login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}